moved directory components to pages
