+++ pages/components/protectedPage.js
